Ticket_NLP_System/requirements.txt
fastapi
uvicorn
scikit-learn==1.6.1
sentence-transformers
faiss-cpu
pandas
numpy